+++ public/index.html
@@ -28,9 +28,12 @@
     <main role="main">
 
         <div class="">
-            <div class="relative px-6 py-32 lg:px-8">
+            <div class="relative px-6 py-2 lg:px-8">
+                <div class="absolute inset-0 z-10">
+                    <img src="./img/pattern-randomized.png" alt="Pattern">
+                </div>
 
-                <div class="mx-auto max-w-2xl py-16 sm:py-32 lg:py-48">
+                <div class="m-auto max-w-2xl py-16 sm:py-24 lg:py-32">
                     <div class="text-center">
                         <h1 class="text-4xl font-medium leading-tight tracking-wide text-gray-900 sm:text-4xl">Hello
                             Typescript 👋 ... 😒 Typical!
@@ -45,15 +48,14 @@
             </div>
         </div>
         <section class="">
-            <div class="py-8 px-4 mx-auto max-w-screen-xl text-center lg:py-16 lg:px-6">
+            <div class="py-2 px-10 mx-auto max-w-screen-xl text-center lg:py-8 lg:px-16">
                 <div class="mx-auto mb-8 max-w-screen-sm lg:mb-16">
                     <h2 class="mb-4 text-2xl tracking-tight font-extrabold text-gray-900 ">Messages</h2>
-                    <p class="font-light text-gray-500 sm:text-xl ">Explore the whole collection of
-                        open-source web components and elements built with the utility classes from Tailwind</p>
+                    <p class="font-light text-gray-500 sm:text-xl ">Explore what the machines have to say</p>
                 </div>
                 <div id="messages_component" class="grid gap-8 lg:gap-16 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
                     <div class="m-auto px-4 py-2 text-center text-gray-500 ring-2 ring-red-300 rounded-2xl">
-                        <div class="mx-auto mb-4 bg-gray-500 w-36 h-36 rounded-full"></div>
+                        <div class="mx-auto mb-4 bg-gray-500 w-16 h-16 rounded-full"></div>
                         <h3 class="mb-1 text-2xl font-bold tracking-tight text-gray-900 ">
                             {{title}}
                         </h3>
